Move fixture helper to a separate file
